+++ src/provider/internal/model/department.go
@@ -1,8 +1,0 @@
-package model
-
-type Department struct {
-	ID       string `json:"id"`
-	Name     string `json:"name"`
-	Parent   string `json:"parent"`
-	Leader   string `json:"leader"`
-}